refactor
- moved classes into own headers
- added jumps to node_t
